requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.3
django-nested-admin==4.0.2
django-nested-formset==0.1.4
greenlet==2.0.1
msgpack==1.0.4
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
